hopefully fixed app postgres queries

+++ data/db/build.sql
@@ -1,20 +1,21 @@
 CREATE TABLE IF NOT EXISTS birthdays (
-    UserID integer PRIMARY KEY,
-    GuildID integer,
+    UserID bigint PRIMARY KEY,
+    GuildID bigint,
     date text
 );
 
 CREATE TABLE IF NOT EXISTS channels (
-    GuildID integer PRIMARY KEY,
-    channel integer
+    GuildID bigint PRIMARY KEY,
+    channel bigint
 );
 
-CREATE TABLE IF NOT EXISTS messages (
-    MessageID integer PRIMARY KEY,
-    guild text,
-    channel text,
-    author text,
-    time text,
-    message text,
-    status text
+CREATE TABLE IF NOT EXISTS messages
+(
+    MessageID bigint PRIMARY KEY,
+    guild     text,
+    channel   text,
+    author    text,
+    time      text,
+    message   text,
+    status    text
 );
